requirements.txt
Flask==3.0.0
Flask-Bootstrap==3.3.7.1
Flask-Moment==1.0.5
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
requests==2.32.3
python-dotenv==1.0.1
WTForms==3.1.2
SQLAlchemy==2.0.32